Final touch ajudstments for bidirectional tool activation
